+++ src/react.spec.ts
@@ -8,7 +8,7 @@ const windowMock = {} as typeof window;
 const windowMockFn = () => windowMock;
 const throwingFn = _.EnvironmentError.throwingFn.bind(null);
 const warningFn = _.EnvironmentError.warningFn.bind(null);
-let consoleWarnSpy: jest.SpyInstance<void, Parameters<Console['warn']>, unknown>;
+let consoleErrorSpy: jest.SpyInstance<void, Parameters<Console['error']>, unknown>;
 
 jest.mock('react', () => {
   function useEffect(effect: React.EffectCallback) {
@@ -23,7 +23,7 @@ jest.mock('react', () => {
 });
 
 beforeAll(() => {
-  consoleWarnSpy = jest.spyOn(console, 'warn');
+  consoleErrorSpy = jest.spyOn(console, 'error');
 });
 
 afterEach(() => {
@@ -44,7 +44,7 @@ describe('Check React utils', () => {
 
   it('useBrowserCallback alikes return fallback function', () => {
     jest.replaceProperty(global, 'window', void 0);
-    consoleWarnSpy.mockImplementation();
+    consoleErrorSpy.mockImplementation();
     const throwingFnSpy = jest.spyOn(_.EnvironmentError, 'throwingFn').mockImplementation(throwingFn);
     const warningFnSpy = jest.spyOn(_.EnvironmentError, 'warningFn').mockImplementation(warningFn);
     expect(_.useBrowserCallback(fn0, []).name).toBe('voidFn');
@@ -63,10 +63,10 @@ describe('Check React utils', () => {
 
   it('useBrowserCallbackOrWarn warning function sends warning', () => {
     jest.replaceProperty(global, 'window', void 0);
-    consoleWarnSpy.mockImplementation();
+    consoleErrorSpy.mockImplementation();
     expect(_.useBrowserCallbackOrWarn(fn0, [])()).toBe(void 0);
-    expect(consoleWarnSpy).toHaveBeenCalledTimes(1);
-    expect(consoleWarnSpy).toHaveBeenLastCalledWith(expect.any(_.EnvironmentError));
+    expect(consoleErrorSpy).toHaveBeenCalledTimes(1);
+    expect(consoleErrorSpy).toHaveBeenLastCalledWith(expect.any(_.EnvironmentError));
   });
 
   it('useBrowserCallbackOrThrow throwing function throws', () => {